Improvements 2
Added an utils file with the auxiliar methods
Reordered imports
Replaced for variables with _ if they aren't used
